+++ common/src/main/java/tocraft/remorphed/command/RemorphedCommand.java
@@ -271,7 +271,7 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
 
     private static void clearShapes(@NotNull CommandSourceStack source, ServerPlayer player) {
         PlayerMorph.getUnlockedShapes(player).clear();
-        PlayerMorph.clearCounter(player);
+        PlayerMorph.getShapeCounter(player).clear();
 
         source.sendSuccess(() -> Component.translatable(Remorphed.MODID + ".clearShapes", player.getDisplayName()), true);
         PlayerShapeChanger.change2ndShape(player, null);
@@ -317,6 +317,7 @@ public class RemorphedCommand implements CommandEvents.CommandRegistration {
 
     private static void clearSkins(@NotNull CommandSourceStack source, ServerPlayer player) {
         PlayerMorph.getUnlockedSkinIds(player).clear();
+        PlayerMorph.getSkinCounter(player).clear();
 
         source.sendSuccess(() -> Component.translatable(Remorphed.MODID + ".clearSkins", player.getDisplayName()), true);
         if (Remorphed.foundSkinShifter) {