+++ common/src/main/java/tocraft/remorphed/impl/RemorphedPlayerDataProvider.java
@@ -28,9 +28,10 @@ public interface RemorphedPlayerDataProvider {
 
     Set<UUID> remorphed$getFavoriteSkins();
 
-    boolean remorphed$handleSwap(ShapeType<? extends LivingEntity> type);
     int remorphed$getCounter(ShapeType<? extends LivingEntity> type);
     int remorphed$getCounter(UUID skinId);
+    void remorphed$handleSwap(ShapeType<? extends LivingEntity> type);
     void remorphed$handleSwap(UUID skinId);
-    void remorphed$clearCounter();
+    Map<ShapeType<?>, Integer> remorphed$getShapeCounter();
+    Map<UUID, Integer> remorphed$getSkinCounter();
 }