+++ common/src/main/java/tocraft/remorphed/screen/widget/SkinWidget.java
@@ -25,8 +25,8 @@ public class SkinWidget extends ShapeWidget {
     private final FakeClientPlayer fakePlayer;
     private final int size;
 
-    public SkinWidget(int x, int y, int width, int height, @NotNull PlayerProfile skin, @NotNull FakeClientPlayer fakePlayer, Screen parent, boolean isFavorite, boolean isCurrent) {
-        super(x, y, width, height, parent, isFavorite, isCurrent);
+    public SkinWidget(int x, int y, int width, int height, @NotNull PlayerProfile skin, @NotNull FakeClientPlayer fakePlayer, Screen parent, boolean isFavorite, boolean isCurrent, int availability) {
+        super(x, y, width, height, parent, isFavorite, isCurrent, availability);
         this.size = (int) (Remorphed.CONFIG.entity_size * (1 / (Math.max(fakePlayer.getBbHeight(), fakePlayer.getBbWidth()))));
         this.skin = skin;
         this.fakePlayer = fakePlayer;