+++ CHANGELOG.md
@@ -2,6 +2,7 @@ remorphed 6.1
 ================
 
 - remove kills after morphing multiple times
+- add a counter to the menu for possible morphs
 - fix crash with special shape widget
 - update to craftedcore 6.0 & walkers 6.2